+++ tests/ObjectReflectorTest.php
@@ -39,7 +39,7 @@ final class ObjectReflectorTest extends TestCase
                 'SebastianBergmann\ObjectReflector\TestFixture\ParentClass::protectedInParent' => 'protected',
                 'SebastianBergmann\ObjectReflector\TestFixture\ParentClass::publicInParent'    => 'public',
             ],
-            $this->objectReflector->getAttributes($o)
+            $this->objectReflector->getProperties($o)
         );
     }
 
@@ -51,7 +51,7 @@ final class ObjectReflectorTest extends TestCase
             [
                 1 => 2,
             ],
-            $this->objectReflector->getAttributes($o)
+            $this->objectReflector->getProperties($o)
         );
     }
 }